+++ NEWS.md
@@ -1,3 +1,12 @@
+# nzilbb.labbcat 0.7-1
+
+Minimum LaBB-CAT version *20210210.2032*
+
+# Enhancements
+
+- remove no.progress parameter (progress bar will appear automatically only in interactive mode).
+- getMatches now supports *overlap.threshold* parameter.
+
 # nzilbb.labbcat 0.6-1
 
 Minimum LaBB-CAT version *20200812.1253*